Added color scheme; Move Homepath component to components/Portraithomepath.js
