+++ src/App.scss
@@ -504,7 +504,8 @@
     }
   }
 
-  $back-colors: ("0", #f9f9f9), ("1", #008000), ("2", #ffffff),
+  $back-colors:("s0",#C490D1),("s1",#ACACDE),
+    ("0", #f9f9f9), ("1", #008000), ("2", #ffffff),
     (
       "2b",
       (linear-gradient(-45deg, rgb(252.5, 252.5, 255), rgb(255, 252.5, 252.5)))
@@ -530,7 +531,8 @@
     ),
     ("23", rgb(250, 250, 250)), ("24", rgb(83, 238, 191));
 
-  $fore-colors: ("0", #008000), ("1", #008000), ("2", #e490b7),
+  $fore-colors: ("s0",#B8336A),
+  ("0", #008000), ("1", #008000), ("2", #e490b7),
     ("4", rgba(56, 56, 56, 0.24)), ("6", rgba(0, 0, 15, 0.34)), ("5", #383838),
     ("7", rgba(56, 56, 56, 0.94)), ("8", #3838383d),
     ("9", rgba(250, 250, 250, 0.94)), ("10", rgb(233, 244, 245)),
